requirements.txt
accelerate==0.29.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cmake==3.29.0.1
contourpy==1.1.1
cycler==0.12.1
dlib==19.24.4
filelock==3.13.3
fonttools==4.53.0
fsspec==2024.3.1
ftfy==6.2.0
huggingface-hub==0.24.0
idna==3.7
imageio==2.34.2
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.4.1
lpips==0.1.4
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
pytorch-fid==0.3.0
pytorch-msssim==1.0.0
pytorch-ssim==0.1
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sympy==1.12
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2023.7.10
tokenizers==0.19.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.42.4
triton==2.2.0
typing_extensions==4.11.0
urllib3==2.2.1
zipp==3.19.2
